+++ backend/bloom/domain/metrics.py
@@ -6,6 +6,19 @@ from enum import Enum
 from bloom.domain.vessel import Vessel,VesselListView
 from bloom.domain.zone import Zone,ZoneListView
 
+class Metrics(BaseModel) :
+    model_config = ConfigDict(arbitrary_types_allowed=True)
+    timestamp : datetime
+    vessel_id: int
+    type : str
+    vessel_mmsi: int
+    ship_name: str
+    vessel_country_iso3: str 
+    vessel_imo: int
+    duration_total : float
+    duration_fishing: Optional[float] = None
+    zone_name : str
+    
 class TotalTimeActivityTypeEnum(str, Enum):
     total_time_at_sea: str = "Total Time at Sea"
     total_time_in_amp: str = "Total Time in AMP"
